chore: split SQL functions in separated files
