+++ internal/management/controller/subscription_controller.go
@@ -18,14 +18,11 @@ package controller
 
 import (
 	"context"
-	"database/sql"
 	"errors"
 	"fmt"
 	"time"
 
 	"github.com/cloudnative-pg/machinery/pkg/log"
-	"github.com/jackc/pgx/v5"
-	"github.com/lib/pq"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
@@ -190,7 +187,7 @@ func NewSubscriptionReconciler(
 	}
 }
 
-// SetupWithManager sets up the controller with the Manager.
+// SetupWithManager sets up the controller with the Manager
 func (r *SubscriptionReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&apiv1.Subscription{}).
@@ -209,8 +206,7 @@ func (r *SubscriptionReconciler) failedReconciliation(
 
 	var statusError *instance.StatusError
 	if errors.As(err, &statusError) {
-		// The body line of the instance manager contain the human
-		// readable error
+		// The body line of the instance manager contain the human-readable error
 		subscription.Status.Error = statusError.Body
 	}
 
@@ -258,164 +254,8 @@ func (r *SubscriptionReconciler) GetCluster(ctx context.Context) (*apiv1.Cluster
 	return &cluster, nil
 }
 
-func (r *SubscriptionReconciler) alignSubscription(
-	ctx context.Context,
-	obj *apiv1.Subscription,
-	connString string,
-) error {
-	db, err := r.instance.ConnectionPool().Connection(obj.Spec.DBName)
-	if err != nil {
-		return fmt.Errorf("while getting DB connection: %w", err)
-	}
-
-	row := db.QueryRowContext(
-		ctx,
-		`
-		SELECT count(*)
-		FROM pg_subscription
-	    WHERE subname = $1
-		`,
-		obj.Spec.Name)
-	if row.Err() != nil {
-		return fmt.Errorf("while getting subscription status: %w", row.Err())
-	}
-
-	var count int
-	if err := row.Scan(&count); err != nil {
-		return fmt.Errorf("while getting subscription status (scan): %w", err)
-	}
-
-	if count > 0 {
-		if err := r.patchSubscription(ctx, db, obj, connString); err != nil {
-			return fmt.Errorf("while patching subscription: %w", err)
-		}
-		return nil
-	}
-
-	if err := r.createSubscription(ctx, db, obj, connString); err != nil {
-		return fmt.Errorf("while creating subscription: %w", err)
-	}
-
-	return nil
-}
-
-func (r *SubscriptionReconciler) patchSubscription(
-	ctx context.Context,
-	db *sql.DB,
-	obj *apiv1.Subscription,
-	connString string,
-) error {
-	sqls := toSubscriptionAlterSQL(obj, connString)
-	for _, sqlQuery := range sqls {
-		if _, err := db.ExecContext(ctx, sqlQuery); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (r *SubscriptionReconciler) createSubscription(
-	ctx context.Context,
-	db *sql.DB,
-	obj *apiv1.Subscription,
-	connString string,
-) error {
-	sqls := toSubscriptionCreateSQL(obj, connString)
-	for _, sqlQuery := range sqls {
-		if _, err := db.ExecContext(ctx, sqlQuery); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func toSubscriptionCreateSQL(obj *apiv1.Subscription, connString string) []string {
-	result := make([]string, 0, 2)
-
-	createQuery := fmt.Sprintf(
-		"CREATE SUBSCRIPTION %s CONNECTION %s PUBLICATION %s",
-		pgx.Identifier{obj.Spec.Name}.Sanitize(),
-		pq.QuoteLiteral(connString),
-		pgx.Identifier{obj.Spec.PublicationName}.Sanitize(),
-	)
-	if len(obj.Spec.Parameters) > 0 {
-		createQuery = fmt.Sprintf("%s WITH (%s)", createQuery, obj.Spec.Parameters)
-	}
-	result = append(result, createQuery)
-
-	if len(obj.Spec.Owner) > 0 {
-		result = append(result,
-			fmt.Sprintf(
-				"ALTER SUBSCRIPTION %s OWNER TO %s",
-				pgx.Identifier{obj.Spec.Name}.Sanitize(),
-				pgx.Identifier{obj.Spec.Owner}.Sanitize(),
-			),
-		)
-	}
-
-	return result
-}
-
-func toSubscriptionAlterSQL(obj *apiv1.Subscription, connString string) []string {
-	result := make([]string, 0, 4)
-
-	setPublicationSQL := fmt.Sprintf(
-		"ALTER SUBSCRIPTION %s SET PUBLICATION %s",
-		pgx.Identifier{obj.Spec.Name}.Sanitize(),
-		pgx.Identifier{obj.Spec.PublicationName}.Sanitize(),
-	)
-
-	setConnStringSQL := fmt.Sprintf(
-		"ALTER SUBSCRIPTION %s CONNECTION %s",
-		pgx.Identifier{obj.Spec.Name}.Sanitize(),
-		pq.QuoteLiteral(connString),
-	)
-	result = append(result, setPublicationSQL, setConnStringSQL)
-
-	if len(obj.Spec.Owner) > 0 {
-		result = append(result,
-			fmt.Sprintf(
-				"ALTER SUBSCRIPTION %s OWNER TO %s",
-				pgx.Identifier{obj.Spec.Name}.Sanitize(),
-				pgx.Identifier{obj.Spec.Owner}.Sanitize(),
-			),
-		)
-	}
-
-	if len(obj.Spec.Parameters) > 0 {
-		result = append(result,
-			fmt.Sprintf(
-				"ALTER SUBSCRIPTION %s SET (%s)",
-				result,
-				obj.Spec.Parameters,
-			),
-		)
-	}
-
-	return result
-}
-
-func (r *SubscriptionReconciler) dropSubscription(ctx context.Context, obj *apiv1.Subscription) error {
-	db, err := r.instance.ConnectionPool().Connection(obj.Spec.DBName)
-	if err != nil {
-		return fmt.Errorf("while getting DB connection: %w", err)
-	}
-
-	if _, err := db.ExecContext(
-		ctx,
-		fmt.Sprintf("DROP SUBSCRIPTION IF EXISTS %s", pgx.Identifier{obj.Spec.Name}.Sanitize()),
-	); err != nil {
-		return fmt.Errorf("while dropping subscription: %w", err)
-	}
-
-	return nil
-}
-
 // getSubscriptionConnectionString gets the connection string to be used to connect to
-// the specified external cluster, while connected to a pod of the specified
-// cluster.
+// the specified external cluster, while connected to a pod of the specified cluster
 func getSubscriptionConnectionString(
 	cluster *apiv1.Cluster,
 	externalClusterName string,
