+++ tests/e2e/declarative_pub_sub_test.go
@@ -1,0 +1,1 @@
+package e2e
